feat | like, review, ask 카운트 Long 클래스로 변경

+++ DutchiePay/src/main/java/dutchiepay/backend/domain/commerce/dto/GetBuyResponseDto.java
@@ -22,10 +22,10 @@ public class GetBuyResponseDto {
     private Integer skeleton;
     private Integer nowCount;
     private LocalDate deadline;
-    private Integer likeCount;
+    private Long likeCount;
     private Boolean isLiked;
-    private Integer reviewCount;
-    private Integer askCount;
+    private Long reviewCount;
+    private Long askCount;
     private Integer[] ratingCount;
     private Double rating;
     private String[] category;
